Change port type to uint16_t to match systemd API.

+++ protocol/include/microsoft/net/remote/protocol/NetRemoteProtocol.hxx
@@ -26,7 +26,7 @@ struct NetRemoteProtocol
     /**
      * @brief Default port if none specified.
      */
-    static constexpr uint32_t PortDefault{ PORT_DEFAULT };
+    static constexpr uint16_t PortDefault{ PORT_DEFAULT };
 
     /**
      * @brief Port separator for address parsing.
